requirements.txt
langchain
langchain-community
langchain-core
streamlit
langchain-groq
transformers
langchain-huggingface
datasets
faiss-cpu
python-dotenv
nltk
